obj json handling + refactoring
